npm: Use npm peerDependencyChecker when no lockfile is present
I've been investigating a slow update scenario for what seems like a
fairly normal scenario. A project with 14 dependencies[1], and no
lockfile present.

From the logs I noticed a ton of requests being made to yarnpkg.com, and
at first I couldn't really place those.

Although there seems nothing weird about this setup, running an update on my
local machine is pretty slow: `7m2.660s`

My laptop is a lot beefier than the average dependabot worker, so it seems like
this update is indeed slow. It's not clear to me at all why that's the case
though, it's a completely normal setup. None of these packages even seem to
need an update, so taking 7 minutes to tell us that on a beefy laptop seems
excessive.

To figure out what's going on, I ran [stackprof](https://github.com/tmm1/stackprof)
on the dryrun for a single update (`axios`) and noticed that a _lot_ of time is
spent in `open3`, which runs our native helpers:

<img width="1014" alt="Screenshot 2021-02-11 at 22 26 05" src="https://user-images.githubusercontent.com/749864/107700907-71905a00-6cb8-11eb-8567-a2959518624b.png">

From tracing the code I noticed that we're running the yarn specific
`yarn:checkPeerDependencies` when we don't have a lockfile. That explains all
the requests to yarnpkg.com. I didn't quite understand why we use the yarn
resolver and it seemed pretty slow, so I replaced it with the
`npm:checkPeerDependencies` when running without a lockfile and the difference
is pretty significant, that runs the entire update in `0m52.091s`.

We used to use the yarn version of this resolver because the npm6
version of this implementation did not handle peer-dependencies
correctly. However, we shipped a new version based on npm7 recently,
and this fixes that problem.

The PR currently defaults any native helper to prefer npm7 over npm6
when there is no lockfile present, but since the `package.json` has not
changed between the two, that seems like a safe change to me.

[1]: `axios, change-case, dayjs, jstz, logrocket, react, react-dom,
react-ga, react-scripts, sass, title-case, babel-eslint,
eslint-config-prettier, prettier`